+++ src/ServiceFabric/ServiceFabric/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 ## Upcoming Release
 
+## Version 3.3.4
+* Fixed secrets exposure in example documentation.
+
 ## Version 3.3.3
 * Updated location of nodeType to use cluster location in stead of resource group location